🩹  NICE-16 Listing (Homepage|Shows|Events) Icon (NEXT-790) [b]
**Listing Rework**:
- [x] Shows => Listing + Upcoming Events
- [x] Events => Listing
- [x] Homepage => Shows Listing

**Icon For Dev**:
- [x] specific import per icon to address: NEXT-790
